refactor(cache): migrate custom endpoints to cachePlugin decorator
Point the 7 custom client endpoints at the relocated publicReadCacheHeaders
(now in @/plugins/cachePlugin) and source their per-endpoint TTLs/tags from
the centralized CUSTOM_READS policy instead of inline literals. Drop the
publicReadCacheHeaders re-export from the @/lib/endpoints barrel and remove
the now-empty src/lib/endpoints/cacheHeaders.ts. Emitted headers are
byte-identical — the existing unit + integration specs pass unchanged
(cache-headers.spec.ts only updates its import path).

Refs #555
