+++ src/collections/Lectures/endpoints/relatedMeditations.ts
@@ -2,12 +2,7 @@ import type { Endpoint, Where } from 'payload'
 
 import { z } from 'zod'
 
-import {
-  commaSeparatedIntIds,
-  parseQuery,
-  publicReadCacheHeaders,
-  requireActiveClient,
-} from '@/lib/endpoints'
+import { commaSeparatedIntIds, parseQuery, requireActiveClient } from '@/lib/endpoints'
 import {
   MEDITATION_CARD_SELECT,
   shapeMeditation,
@@ -15,6 +10,7 @@ import {
 } from '@/lib/meditations/meditationShape'
 import { scoreMeditationByNodes } from '@/lib/meditations/nodeWeights'
 import type { Lecture, Meditation, MeditationsSelect, SubtleSystemNode } from '@/payload-types'
+import { CUSTOM_READS, publicReadCacheHeaders } from '@/plugins/cachePlugin'
 import { asTrustedReq } from '@/plugins/usage/hooks'
 
 /**
@@ -238,10 +234,7 @@ export const lectureRelatedMeditations: Endpoint = {
       relevanceCount === 0 || fallbackShaped.length > 0 ? 'fallback' : 'relevance'
 
     return Response.json({ docs, source, relevanceCount } satisfies RelatedMeditationsResponse, {
-      headers: publicReadCacheHeaders(req, {
-        sMaxAge: 600,
-        tags: ['meditations', 'lectures'],
-      }),
+      headers: publicReadCacheHeaders(req, CUSTOM_READS.lectureRelatedMeditations),
     })
   },
 }
