+++ tests/unit/cache-headers.spec.ts
@@ -4,8 +4,8 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 vi.mock('@/lib/utilities/previewSecret', () => ({ hasValidPreviewSecret: vi.fn() }))
 
-import { publicReadCacheHeaders } from '@/lib/endpoints/cacheHeaders'
 import { hasValidPreviewSecret } from '@/lib/utilities/previewSecret'
+import { publicReadCacheHeaders } from '@/plugins/cachePlugin/cacheHeaders'
 
 // The helper only forwards `req` to hasValidPreviewSecret (mocked here), so a
 // bare stub is enough.
